+++ apps/pleiepenger-sykt-barn/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @navikt/pleiepenger-sykt-barn
 
+## 3.15.6
+
+### Patch Changes
+
+-   Updated dependencies [7f61cd6]
+    -   @navikt/sif-common-soknad-ds@12.0.5
+
 ## 3.15.5
 
 ### Patch Changes